modules: lift make_layout/make_context to BaseConfiguration
